+++ src/main/java/seedu/avo/Main.java
@@ -14,6 +14,7 @@ import seedu.avo.storage.Storage;
 import seedu.avo.storage.TaskParser;
 import seedu.avo.tasks.Task;
 import seedu.avo.tasks.TaskManager;
+import seedu.avo.ui.MainWindow;
 import seedu.avo.ui.ResponseFormatter;
 
 /**
